dev: semana eva 3

+++ v2/pregunta-frecuente/index.php
@@ -88,16 +88,27 @@ if ($_version == 'v2') {
                     include_once 'controller.php';
                     include_once '../conexion.php';
                     $control = new Controlador();
+                    $contadorEntradas = 2;
+
                     $body = json_decode(file_get_contents("php://input", true));
                     // var_dump($body);
                     if (strlen($body->pregunta) > 0) {
                         $respuesta = $control->putPreguntaById($body->pregunta, $body->id);
+                        $contadorEntradas--;
                     }
                     if (strlen($body->respuesta) > 0) {
                         $respuesta = $control->putRespuestaById($body->respuesta, $body->id);
+                        $contadorEntradas--;
                     }
-                    http_response_code(200);
-                    echo json_encode(["data" => $respuesta]);
+                    if ($contadorEntradas <= 1) {
+                        //entro y modifico las 2
+                        http_response_code(200);
+                        echo json_encode(["data" => $respuesta]);
+                    } else {
+                        //no entro ni modifico
+                        http_response_code(409);
+                        echo json_encode(["data" => false]);
+                    }
                 } else {
                     http_response_code(401);
                     echo json_encode(["Error" => "No tiene autorizacion PUT"]);
@@ -109,11 +120,24 @@ if ($_version == 'v2') {
                     include_once '../conexion.php';
                     $control = new Controlador();
                     if ($existeId) {
-                        $respuesta = $control->deleteById($valorId);
-                        http_response_code(200);
-                        echo json_encode(["data" => $respuesta]);
+                        $listado = $control->getAll();
+                        $existeRegistro = false;
+                        foreach ($listado as $item) {
+                            if ($item['id'] == $valorId) {
+                                $existeRegistro = true;
+                            }
+                        }
+                        if (!$existeRegistro) {
+                            http_response_code(404);
+                            echo json_encode(["data" => false]);
+                        } else {
+                            $respuesta = $control->deleteById($valorId);
+                            http_response_code(200);
+                            echo json_encode(["data" => $respuesta]);
+                        }
                     } else {
-                        echo 'error faltan parametros';
+                        http_response_code(406);
+                        echo json_encode(["data" => false]);
                     }
                 } else {
                     http_response_code(401);
